Move admonition code to a separate file
This is intended to help clean things up and isolate the admonition
support in a distinct place.
